backend/requirements.txt
Flask>=2.2
tensorflow>=2.11
numpy
pillow
python-dotenv
openai
gunicorn
flask-cors
matplotlib
requests